Move sharding statement to sharding module
