+++ shardingsphere-features/shardingsphere-sharding/shardingsphere-sharding-distsql/src/main/java/org/apache/shardingsphere/sharding/distsql/parser/statement/DropShardingTableRuleStatement.java
@@ -15,17 +15,21 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.distsql.parser.statement.rql.show.impl;
+package org.apache.shardingsphere.sharding.distsql.parser.statement;
 
-import org.apache.shardingsphere.distsql.parser.statement.rql.show.ShowRulesStatement;
-import org.apache.shardingsphere.sql.parser.sql.common.segment.generic.SchemaSegment;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.apache.shardingsphere.distsql.parser.statement.rdl.drop.DropRuleStatement;
+import org.apache.shardingsphere.sql.parser.sql.common.segment.generic.table.TableNameSegment;
+
+import java.util.Collection;
 
 /**
- * Show sharding binding table rules statement.
+ * Drop sharding table rule statement.
  */
-public final class ShowShardingBindingTableRulesStatement extends ShowRulesStatement {
+@RequiredArgsConstructor
+@Getter
+public final class DropShardingTableRuleStatement extends DropRuleStatement {
     
-    public ShowShardingBindingTableRulesStatement(final SchemaSegment schema) {
-        super(schema);
-    }
+    private final Collection<TableNameSegment> tableNames;
 }
